scrub-patch: clarify renames, and no longer strip related
Stripping these doesn't matter given once the "diff --git" part is
removed (which we /usually/ want gone), patch(1) will not do renames.
No sense in cleaning further bits from a patch that's broken.

Also clarify QA message, while it's possible to use these it doesn't
seem like a great idea with portage. If these are used in the tree,
it'll be dangerous to scrub given patch will still apply but be wrong.