qa-vdb.exclude-lib: initial import to exclude libraries
Some non-gcc packages (like gnat-gpl and julia) provide libgcc_s.so
and similar libraries, and then when a package has it in DT_NEEDED
it looks for a potential provider and believes these might be one.

Optimally would want to auto-exclude packages' private libraries,
but it is difficult to detect (especially if rpath or non-standard
ldpath gets used), so instead providing a config file that allow
to ignore certain libraries from DT_NEEDED.

Alternatively, gnat-gpl and julia could be added to qa-vdb.exclude
but this is to provide a more general mechanism.

Still keeping /gcc in .exclude for simpler exclusion.

Ultimately users are encouraged to set their own excludes based
on their system, installed packages, and toolchain settings.

Fixes #3