bashrc: add guard for pkgcore
This is both to avoid spurious messages when happen to use pmerge
to test things, and because undefined behavior from the likes of
qa-sed may even break merging packages.

Only adding this to the example bashrc and so, if someone really must
use it, can use a custom bashrc and there is no checks anywhere else.