qa-cmp: try to improve tocompare ordering for system
At the base it compare oldest image with newest, but if specify exact
=cat/pkg-ver(newest) with =cat/pkg-ver(oldest) (i.e specifically
request to compare in reverse) and one of these is a system copy,
it'd ignore ordering.

This should hopefully(?) fix that without regressions.

This whole function could use a rewrite, can hardly make sense of
it anymore and it was the biggest issue in writting qa-cmp with
the goal of allowing non-specific arguments to save users' time.