scripts: add option to set EPREFIX
Most Gentoo Prefix issues were already fixed while trying to
support ROOT (for tests), just need to actually know EPREFIX.

Getting value of EPREFIX is a bit annoying, given it's not
exported in prefix environments, the normal way would be to call
`portageq envvar EPREFIX` but portageq has slow startup and don't
want to check for this for every non-prefix users.

So opting to add a -Deprefix option that the ebuild can use for
a default (users can still set it in the env manually).

Using ': @eprefix@' so the tool is usable even if wasn't configured.

Also normalize use of ROOT/EROOT/EPREFIX.

There may still be more that needs to be handled but it seems to work.