scripts: setup stronger || die
This being annoying with bash is nothing new, but this
should improve things a bit.

Few elements:
  ERR: propagated trap for all nested subshells to die /immediately/
/stat: get real ppid of current bashpid (this isn't always $PPID),
       this allows the annoying `function $(die)` to terminate
       as well without going full `kill 0` with its own issues
 USR1: to send 166 to real ppid, or common 1 if main process
 USR2: if main process needs to return 0

May eventually offload this to a shared include file in non-compact
form, but will wait till there's more code to share (e.g. possibly
bash-based qatom as it's inconvenient to rely on external).