scripts: various help text / man pages improvements
man pages now have a short header text to describe the command that's
set in scripts/meson.build

bashrc documentation was moved out of .bashrc files inside the --help
text and will also be present in man pages

Parts with asterisks (e.g. *Notes*) is for help2man to create a
section. Not the prettiest output but, until decide to use something
a proper system for docs that'd also handle --help, that'll have to do.