qa-cmp: slightly improve filelist version replacement, add options
Figure not everyone may want this, so keeping it default but adding
options to either reduce the behavior or disable entirely.

Also try to strip _* parts from versions which are rarely used
outside ebuild names. Not that can pickup if 1.0_rc1 was converted
to 1.0-rc1 and is actually being used.

Using '<snip>' to hopefully make intention clearer than '${PV}'