qa-cmp: refactor image2atomf handling
Early stages qa-cmp could work only on the image/ directory itself,
but upon further work started to rely on the build-info/ directory.

Rather than rely on path to find the version string, use build-info
as well. This is an alternative fix to adding the missing set +f for
the "hack" that'd check usr/share/doc/* to find the version string.

This allows to move/backup ${PORTAGE_TMPDIR}/category/<dir> to e.g.
~/mydir and `qa-cmp mydir mynewdir` with cleaner output.