common.bashlib: remove kill
Better way to carry over an exit code from a subshell, but requires
non-masked return values from anything that may call die/end.

This means the test "declare v=$(echo $(false || die))" where expected
to exit can no longer work, but idea is to not use this kind of
statement in the first place. Relying on kill is just a bad thing.