qa-vdb: migrate to common/atomf.bashlib
Not particularly happy with the state of qa-vdb and feel it needs
a rewrite. But for now will just bring it in-line even if it means
not-nice hacks.

Ultimately idea would be to have a vdb.bashlib for general purpose
vdb access that'd be usable by both -vdb and -cmp.