scripts: replace QA: in message headers
Replaced by CMP;, VDB:, or SED:

Avoids redundancy if portage goes forward with using [QA] or
similar other prefixes to distinguish, and tells us which
check the message is coming from.

Originally thought to remove the 'QA: ' prefix entirely, but
it makes it difficult to distinguish where a new QA notice
block starts, so gave it a renewed use.

Unchanged in scrub-patch given not an header nor used by portage.