qa-vdb: add workaround for ebuilds with pipewire+jack
qa-vdb tries to not nag about e.g.
  -virtual/jack (doesn't install anything, useless dep)
  +media-sound/jack2 (has the library)

This also works for pipewire being a libjack provider, however if
ebuild has virtual/jack and pipewire at same time, we get:
  -media-sound/pipewire:=
Believing it's provided by virtual/jack.

This workarounds this for at least pipewire given the two being
together is typical and will be more frequent for audio applications.
A real solution would prove difficult, and currently this does not
seem to be a widespread issue.