scripts: fix portageq during python migration
A world upgrade doing "PYTHON_TARGETS=python3_9 -> python3_10" without
keeping 3_9 leads constant failure in qa-vdb/qa-cmp for every packages
after portage was upgraded until a new instance of emerge:
 * qa-vdb: running 'qa-vdb' failed (disable with QA_VDB=n)
 * /usr/lib/python-exec/python3.9/portageq: this Python implementation
   (python3.9) is not supported by the script.
 * Error: portageq failed to return vdb path (qa-vdb:vdb-set_vdb:680)

This is due to attempting to run ebuild-helpers/portageq wrapper that
the running instance of portage added to PATH.

Running portage's wrapper script is pointless here, so strip from PATH
as a workaround. Prefer to avoid absolute path and still rely on the
normal path to find it (just not ebuild-helpers/).

At the same time, merge the `unset DEBUG` for q* workaround in a
common function for these workarounds.