atomf.bashlib: workaround strange bash behavior on non-Linux
For convenience, atomf.bashlib uses regexes that match on nothing (''),
e.g.
    val="app-portage/iwdevtools-0.9.0"
    [[ ${val} =~ ^('!!'|'!'|'')(.*) ]] && echo a || echo b
so BASH_REMATCH[1] will be !!, !, or lastly nothing, and [2] the rest,
meaning [1]="" [2]="app" and echos "a".

For that particular line, it shouldn't even be possible to echo b
(matching on nothing + any characters or none, so string can be empty),
and yet on a macOS Gentoo prefix with bash-5.1_p16.. it did.

Unsure which one is intended behavior given matching nothing is a bit
unconventional -- but by giving it a rough equivalent with .{0}
(meaning match any characters 0 times), it works as expected.

e.g.
    [[ ${val} =~ ^('!!'|'!'|.{0})(.*) ]] && echo a || echo b

Thanks-to: Sam James <sam@gentoo.org>