qa-cmp: support comparing file permissions
Given VDB doesn't store file permissions, it's required to check
filesystems directly and there may be disparity between system's
and images. Given previous issues when tried to check sizes directly
this is likely to cause some more, so have some degree of failsafes
and --ignore-perms as last resort.

pkg_postinst() changes will also cause a disparity, but if this
happens everytime rather than 1-time fix there's probably an issue
with that postinst.

Needed a workaround for FEATURES=sfperms given the permissions
aren't changed on the image but on merge and can't compare. Not
much can do if someone uses FEATURES=suidctl or other fancy things.