scripts: fix usage when files have $ in them
Wasn't aware of this, but [[ -v array[${var}] ]] seems to work like
eval, meaning that if ${var} has $ in it, it tries to read the variable
and fails on unbound error. Using ["${var}"] doesn't help.

Undocumented and feel may possibly be a bash bug?

[[ ${array[${var}]+x} ]] is equivalent and doesn't cause this, so
let's use it where it's risky.

Closes #7