portage: drop pkgcore guard in example bashrc
This is meant to be a simple example file and don't want more
checks than necessary in there.

Could move check to individual *.bashrc, but this does not seem overly
relevant as these are advertised for /portage/ integration in the
first place. Not a fan of pkgcore trying to use /etc/portage, it
also behave differently for post_pkg_* given it only calls them if
was defined in the ebuild).

At worst pmerge will give some error messages but not break anything.