qa-openrc: use shellimport
Just a drop-in replacement to ensure not breaking anything, may make
sense to set SI globally so that multiple checks re-use it to check
different variables and functions (but this just checks ssd right now).

If a init.d script decides to `rm -rf /${RC_EMPTYVAR}` in global scope
for some reason it now won't delete anything even without sandbox.