scripts: fix with bash-5.2
Go back to the original +x version which works with both 5.1 and 5.2.
Preferred -v given it's the bash way to check if something is set, but
until depend on >=bash-5.2 that's not going to work.

To recap what's happening, take:

    index='$(echo test >&2)'
    declare -A array
    array['$(echo test >&2)']=set

bash-5.1:
    [[ -v array[${index}] ]] && echo set # will echo "test"
(original issue was a path having a $ in it being interpreted)
    [[ -v array['${index}'] ]] && echo set # will echo "set"
bash-5.2:
    [[ -v array[${index}] ]] && echo set # will echo "set" # good
    [[ -v array['${index}'] ]] && echo set # nothing

The quotes workaround was a stretch, so not surprising it broke.
So instead use: [[ ${array[${index}]+x} ]]