qa-vdb: warn about qml on removal suggestion
Kind of a bad hack, but this comes up often and removal suggestions
may mislead maintainers that are not used to work with QT.

Currently does not check if the suggested removals do provide /qml/
files as it'd need to scan CONTENTS rather than just PROVIDES and may
not be worth it. Will be quiet if the package doesn't use qtdeclarative
in which case suggestions are probably correct.

Giving adequate qml suggestions would require a separate tool as VDB
doesn't have the required information. Could potentially update the
warning in the future to point to a way to verify.