qa-sed.bashrc: escape backslashes for portage log command
portage uses `echo -e` which converts \r and can get messy with
e.g. dos2unix sed lines, so escape backslashes before feeding to
eqawarn (shouldn't be an issue with non-sed bashrc).

Small downside is that this assumes QA_SED_LOG's command need this
(may not even be from portage).

There is also a quirk to handle:

This works with >=bash-5.0: `${var//\\/\\\\}` or `${var//'\'/'\\'}`
...but doesn't with bash-4.2 or if compat42 is set, and instead
need this messy: `${var//\\/\\\\\\}` ('\\' doesn't work).

Unsure for universal way to do this without calling external commands
or escaping other things (like with %q), so do a EAPI Check.