repo-cd: advertise eshowkw further in --help (#13)
An example stands out more than the block of text.

Using `eshowkw` over `pkgdev showkw` because the latter generates
metadata cache which may not always be desirable when just changing
directory. `pkgdev showkw` is however faster than `eshowkw`.

Hard to decide on default as it mostly comes down to preferences,
and rather leave that up to users if want more than `ls -1v`.