portage: ensure ROOT and EPREFIX are exported
Export behavior could potentially change in portage in the
future and doesn't seem overly worth implementing arguments
to pass it (could reconsider if felt useful), so just prefix
to the call.

Not really used by all of those (e.g. sed), but keep same
across to maintain expectations.

EROOT is not needed given common.bashlib will concat ROOT and
EPREFIX as-is, but it'd be used if happens to be exported.