qa-cmp: be bit more strict with SLOT replacement
Unlike versions, SLOT does not necessarily start with a number,
and if we had e.g. SLOT=e, then "somefile" would become "som*fil*".
So further check that it is not near letters on top of numbers.

PMS states:
  A slot name may contain any of the characters [A-Za-z0-9+_.-].
  It must not begin with a hyphen, a dot or a plus sign.

So thankfully do not have to worry about [+.-], but ignore
_ given SLOT="_something" is rather unusual and _<slot> is
reasonable to replace.