qa-cmp: normalize symlink permissions in files lists
Not quite sure when this started, but lately started to see symlinks
with lrwxr-xr-x (With ls or stat) on ext4, then mv or cp -a them to
a tmpfs and they're now the lrwxrwxrwx that more used to.

This causes noise when comparing, so let's normalize it.

(Also shellcheck, please, this is not a multidimensional array...
Would've used l????????? but given this hackily matches the whole
string this could pickup a ` (lword ` in filenames)