tests: work around portage confusion by passing dev-test/
In theory portage should be ignoring ::gentoo (different repos.conf
which does not specify /var/db/repos/gentoo and changes main-repo),
but seems it gets confused by a package named "test" existing still
resulting in:

    You specified an unqualified atom that matched multiple packages:
    * dev-test/test
    * dev-python/test

Whether category is passed or not is not important in env-atom and
*-perms (so can just do it), but in "qa-cmp/pick" it was omitted on
a few tests to test the missing category logic.

So we lose that extra bit until can figure out a way for portage to
actually ignore ::gentoo. Not passing the category is however rarely
done given the portage bashrc hook passes it, this is for when using
qa-cmp directly and being brief with arguments.