scrub-patch: try harder to check if a unified patch
file-5.45 was reporting mbox patches (e.g. if fetch a <commit>.patch
from github) as a unified format patch suitable for the tree, but
file-5.46 only says they're just an mbox file now.

Not ideal, but use -k to check what the 2nd best match is.

Also toss LC_ALL while at it just in case, ideally would use -i but
we want to know if it's unified or not rather than just a "x-diff".