tests: fix scrub-patch tests with file-5.46
Regardless of the previous commit making file try harder,
file-5.46 is now quite insistent that mess.patch is actually
a unified patch on the first match ignoring #!/bin/sh.

So use something a bit stronger to convince it that it's not.

The previous commit change will at most try 2 matches, so need
a 2nd "bad" match as well. In this case it is first a Gentoo
metadata.xml, or a normal XML document.

Wanted to have a 3rd bad choice but file(1) is quite insistent
that it is a diff unless dump a large amount of data so it
gives up.