tests: add note about PM spec validity of two atoms
Just for clarity given the test is sometime looked at by people
implementing parsers.

"Believe" it is not valid, but PM spec may leave a bit of room for
interpretation there, e.g. does "empty" count as a "slot dependency"
that can be followed by an optional sub-slot like ':/0'? Does '*'
only has a meaning for matching slots and not sub-slot? '*/sub' to
match any slot that has a sub-slot of 'sub' could potentially make
sense but may be bit of a stretch.

Either way, iwdevtools is intended to accept incomplete atoms (e.g.
missing category for lazily given package atoms on the command line,
this could include slots without subslot as well) and is not a validity
tester for actual ebuild DEPEND strings. So the test itself is fine.