qa-sed: migrate function == phase check to bashrc
Debated a few ways to fix this, like passing different arguments,
or exporting a variable in bashrc... but ultimately feels that the
check itself belongs in bashrc. As a bonus it'll reduce the amount
of external calls to qa-sed too (aka eclasses' sed will call sed
directly).

On the downside, if we ever want to make this code more complex
to allow functions in ebuilds (not eclasses), then that will be
messier to have in the bashrc itself as these are meant to be
small/simple.

phase-only tests were also migrated to the bashrc tests